Refactor enemy attack functions
Attacks on the same fields kept happening, which
stopped the attack function from executing.
Reason was due to the race condition between
simulateClickAndObserve and the promise / animation
handling of the attack flow.

Now safety checks have been installed to avoid double
attacks and in case of some it will handle them as
missed to let the gameflow continue.